Fixed a unit test and updated the logging of HttpStatusCode to log the int instead of the enum string

+++ ConsulRx/MonitoringExtensions.cs
@@ -8,7 +8,7 @@ namespace ConsulRx
     {
         public static void IncludeConsulResult<T>(this EventContext eventContext, QueryResult<T> result)
         {
-            eventContext["HttpStatusCode"] = result.StatusCode;
+            eventContext["HttpStatusCode"] = (int)result.StatusCode;
             eventContext["ResponseIndex"] = result.LastIndex;
             eventContext["KnownLeader"] = result.KnownLeader;
